Update Bcrypt.pas
C++Builder does not need to see the definition of CP_UTF8. The EXTERNALSYM directive tells the compiler not to generated code for this symbol in Bcrypt.hpp.

The way it is defined causes an error as wel with the 32-bit Clang compiler: [C++ Error] Bcrypt.hpp(132, 34):  expected unqualified-id. Not generating the definition makes this a non-issue. 

I only added a single line but something in the text makes it look like 12 lines were changed.